stome: rename file/dir in-place by F2

+++ frontend/src/stome/Content.js
@@ -6,6 +6,9 @@ import ContentPanel from './ContentPanel'
 export default class Content extends React.Component {
   constructor(props) {
     super(props);
+    this.state = {
+      renaming: false,
+    };
     window.on('keyup', this.onKeyUp);
   }
 
@@ -21,6 +24,8 @@ export default class Content extends React.Component {
           files={this.props.currentDir.files}
           onClick={this.props.onClick}
           onMouseDown={(ev) => ev.preventDefault()}
+          onStopRename={this.onStopRename}
+          renaming={this.state.renaming}
         />
         <ContentPanel
           item={this.props.currentItem}
@@ -33,14 +38,39 @@ export default class Content extends React.Component {
   }
 
   onKeyUp = (ev) => {
-    if (ev.key === 'Delete') {
-      this.deleteItem(this.props.currentItem);
+    switch (ev.key) {
+      case 'Delete':
+        this.deleteItem(this.props.currentItem);
+        break;
+      case 'F2':
+        this.renameItem(this.props.currentItem);
+        break;
+      default:
+        break;
     }
   }
 
   deleteItem = async (item) => {
+    if (!item) return;
     await item.delete();
     this.props.onItemDeleted(item);
     this.props.onItemChange();
+  }
+
+  renameItem = async (item) => {
+    if (!item) return;
+    this.setState({
+      renaming: true,
+    });
+  }
+
+  stopRename = () => {
+    this.setState({
+      renaming: false,
+    });
+  }
+
+  onStopRename = () => {
+    this.stopRename();
   }
 }
